# tz: optimize AVX2 implementation